+++ libraries/VarAPI/include/var/Set.hpp
@@ -59,21 +59,14 @@ private:
 
 };
 
-#if 0
-//TODO No reverse iterator. Can't just inherit ContainerObject
 template <typename Key, class Hash = std::hash<Key>>
-class UnorderedSet : public ContainerObject<UnorderedSet<Key,Hash>, std::unordered_set<Key, Hash>, Key> {
-  using Base = ContainerObject<Vector<Key>, std::unordered_set<Key, Hash>, Key>;
+class UnorderedSet : public ContainerObjectForwardOnly<UnorderedSet<Key,Hash>, std::unordered_set<Key, Hash>, Key> {
+  using Base = ContainerObjectForwardOnly<Vector<Key>, std::unordered_set<Key, Hash>, Key>;
 
 public:
-
   using Self = UnorderedSet<Key,Hash>;
-
   using iterator = typename std::unordered_set<Key,Hash>::iterator;
   using const_iterator = typename std::unordered_set<Key,Hash>::const_iterator;
-  using reverse_iterator = typename std::unordered_set<Key,Hash>::reverse_iterator;
-  using const_reverse_iterator =
-    typename std::unordered_set<Key,Hash>::const_reverse_iterator;
 
   iterator insert(const_iterator hint, const Key &value) {
     return this->m_container.insert(hint, value);
@@ -105,7 +98,91 @@ public:
 private:
 
 };
-#endif
+
+template <typename Key, class Compare = std::less<Key>>
+class OrderedMultiset : public ContainerObject<OrderedSet<Key,Compare>, std::multiset<Key, Compare>, Key> {
+  using Base = ContainerObject<Vector<Key>, std::multiset<Key, Compare>, Key>;
+
+public:
+
+  using Self = OrderedSet<Key,Compare>;
+
+  using iterator = typename std::multiset<Key,Compare>::iterator;
+  using const_iterator = typename std::multiset<Key,Compare>::const_iterator;
+  using reverse_iterator = typename std::multiset<Key,Compare>::reverse_iterator;
+  using const_reverse_iterator =
+    typename std::multiset<Key,Compare>::const_reverse_iterator;
+
+  iterator insert(const_iterator hint, const Key &value) {
+    return this->m_container.insert(hint, value);
+  }
+
+  Self& insert(const Key & value){
+    this->m_container.insert(value);
+    return *this;
+  }
+
+  Self& insert(Key && value){
+    this->m_container.insert(std::forward<Key>(value));
+    return *this;
+  }
+
+  Self& erase(const Key & value){
+    this->m_container.erase(value);
+    return *this;
+  }
+
+  iterator lookup(const Key & value){
+    return this->m_container.find(value);
+  }
+
+  const_iterator lookup(const Key & value) const {
+    return this->m_container.find(value);
+  }
+
+};
+
+template <typename Key, class Hash = std::hash<Key>>
+class UnorderedMultiset : public ContainerObjectForwardOnly<UnorderedSet<Key,Hash>, std::unordered_multiset<Key, Hash>, Key> {
+  using Base = ContainerObjectForwardOnly<Vector<Key>, std::unordered_multiset<Key, Hash>, Key>;
+
+public:
+
+  using Self = UnorderedMultiset<Key,Hash>;
+
+  using iterator = typename std::unordered_multiset<Key,Hash>::iterator;
+  using const_iterator = typename std::unordered_multiset<Key,Hash>::const_iterator;
+
+  iterator insert(const_iterator hint, const Key &value) {
+    return this->m_container.insert(hint, value);
+  }
+
+  Self& insert(const Key & value){
+    this->m_container.insert(value);
+    return *this;
+  }
+
+  Self& insert(Key && value){
+    this->m_container.insert(std::forward<Key>(value));
+    return *this;
+  }
+
+  Self& erase(const Key & value){
+    this->m_container.erase(value);
+    return *this;
+  }
+
+  iterator lookup(const Key & value){
+    return this->m_container.find(value);
+  }
+
+  const_iterator lookup(const Key & value) const {
+    return this->m_container.find(value);
+  }
+
+};
+
+
 
 } // namespace var
 #endif // VAR_API_SET_HPP_